LPD-61987 move and improve test
